KEYS: Add a new keyctl op to reject a key with a specified error code
Patchwork-id: 45798
O-Subject: [RHEL6 PATCH 19/38] KEYS: Add a new keyctl op to reject a key with a
	specified error code
Bugzilla: 788634
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Add a new keyctl op to reject a key with a specified error code.  This works
much the same as negating a key, and so keyctl_negate_key() is made a special
case of keyctl_reject_key().  The difference is that keyctl_negate_key()
selects ENOKEY as the error to be reported.

Typically the key would be rejected with EKEYEXPIRED, EKEYREVOKED or
EKEYREJECTED, but this is not mandatory.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: James Morris <jmorris@namei.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>