KEYS: Make the key reaper non-reentrant
Patchwork-id: 45808
O-Subject: [RHEL6 PATCH 29/38] KEYS: Make the key reaper non-reentrant
Bugzilla: 788634
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Make the key reaper non-reentrant by sticking it on the appropriate system work
queue when we queue it.  This will allow it to have global state and drop
locks.  It should probably be non-reentrant already as it may spend a long time
holding the key serial spinlock, and so multiple entrants can spend long
periods of time just sitting there spinning, waiting to get the lock.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: James Morris <jmorris@namei.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>