KEYS: Add missing smp_rmb() primitives to the keyring search code
Patchwork-id: 45812
O-Subject: [RHEL6 PATCH 33/38] KEYS: Add missing smp_rmb() primitives to the
	keyring search code
Bugzilla: 788634
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Add missing smp_rmb() primitives to the keyring search code.

When keyring payloads are appended to without replacement (thus using up spare
slots in the key pointer array), an smp_wmb() is issued between the pointer
assignment and the increment of the key count (nkeys).

There should be corresponding read barriers between the read of nkeys and
dereferences of keys[n] when n is dependent on the value of nkeys.

Signed-off-by: David Howells <dhowells@redhat.com>
Reviewed-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: James Morris <jmorris@namei.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>