kdump: round up total_size to 128M for crashkernel reserving threshold
Patchwork-id: 45838
O-Subject: [RHEL6.3 PATCH v2]kdump: round up total_size to 128M for crashkernel
	reserving threshold
Bugzilla: 798727
RH-Acked-by: Amerigo Wang <amwang@redhat.com>

[v1 -> v2]: Per Amerigo Wang round up the total size to 128M instead of use
2G-128M as threshold.

In bug 772311 KEXEC_AUTO_THRESHOLD was lowered to 2G (1ULL<<31) for x86_64 and
i686, however during QA testing it was discovered that systems w/ exactly 2G
ram will not have crashkernel memory reserved.

The total memory size we get in kernel is usually slightly less than 2G with a
2G memory module machine [one exeption is kvm guest run with native kvm tool].
The reason of this issue is that bios will reserve some area for its own use
and it will not export all 2G memory as usable in e820.

An option is to use dmi/smbios to get physical memory size, but it's not
reliable as well. As Prarit said hardware vendors sometimes will screw this up.
Thus we choose to round up total size to 128M to workaround this problem.
It should cover the 2G machines.

OTOH, there's another bug with crashkernel reserve get_total() function. It uses
max_low_pfn - min_low_pfn to get the total memsize which is wrong because it
will including the memory holes bios advertised. I sent a patch to upstream:
https://lkml.org/lkml/2012/3/9/690 they are two problems related but due to
we have the 2G threshold commited already so we can resolve this firstly
and backport that patch to rhel later.

Upstream status:
This is not an upstream problem.

Resolves: bug 798727
Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=798727

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=4150687

Test:
Tested on 2G kvm guest (qemu-system-x86_64) and below beaker machine:
dell-p390n-01.lab.bos.redhat.com

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>