mlx4: Don't show RoCE interfaces if the hpn channel is not installed
Patchwork-id: 45842
O-Subject: [Patch RHEL6] mlx4: Don't show RoCE interfaces if the hpn channel is
	not installed
Bugzilla: 753004
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

The HPN channel is where we officially support RoCE/IBoE interfaces.  When
we first created it, the user space packages in the base channel did not
include the RoCE/IBoE patches, however the kernel did because we don't
have separate kernels for the HPN channel and the base channel.  This
created confusion where the user space didn't know about RoCE/IBoE
devices and the kernel did.  In addition, people were attempting to use
the RoCE devices without the HPN add on, not because they were trying
to get around buying the HPN add on, but because they didn't know it was
even needed (and that's totally understandable when the devices showed up
without it).  So, this patch makes the devices disappear unless the HPN
channel is added on to the base subscription.

Note: this is a "keep people honest" type change, it is not intended to
be something people can't work around.  When people see that their RoCE
devices don't show up by default in the base product, then they'll ask
what they need to do to enable them, the answer will be "buy the HPN
add on", and there will no longer be any confusion.  However, only a
modest amount of digging would uncover that they can set the hpn
variable themselves, so this isn't intended to be an enforcement mechanism,
just a reduction of confusion mechanism.

Tested on my cluster.  Without the libmlx4-rocee package and with this
patch, all RoCE devices show up as plain 10GigE devices and no more using
any of ibv_devices, ibv_devinfo, ibstat, and ibstatus (these check
different paths to the device info, and they all consistently do not
show the RoCE device as RDMA capable).  With the libmlx4-rocee package
in place which includes a modprobe.d/libmlx4.conf file that sets hpn
to 1, the RoCE devices come right back and are seen via all of the
above listed mechanisms.

Resolves: bz753004

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>