mpt2sas: reduce lock contention
Patchwork-id: 45863
O-Subject: [RHEL6.4 PATCH] mpt2sas: reduce lock contention
Bugzilla: 790506
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz 790506

The writeq_lock is needed for 32bit in order to ensure that the writes from the drivers
don't get intertwined. This is however not needed on 64bit systems where the writeq
is atomic. With this patch, when writeq and CONFIG_64BIT is defined the driver uses
the system provided writeq otherwise it uses a locking mechanism and two writel calls.
Reporter tested a similar patch and has seen a performance improvement:
"On a 4S WSM-EX system, we saw a 1% performance improvement on our OLTP workload"

The patch has been acked by the LSI maintainer.

Upstream - the writeq is defined only for systems where the writeq is atomic, RH6 has
it defined also for 32bit systems. That is why our code differs from upstream.

Brew - taskID=4158965

Tomas

Signed-off-by: Jarod Wilson <jarod@redhat.com>