GFS2: Remove a __GFP_NOFAIL allocation
Patchwork-id: 45948
O-Subject: [RHEL6.3 PATCH BZ 796017 2/3] GFS2: Remove a __GFP_NOFAIL allocation
Bugzilla: 796017
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>
Date: Fri, 9 Mar 2012 13:14:44 -0600
Subject: [PATCH 2/3] Subject: GFS2: Remove a __GFP_NOFAIL allocation

In order to ensure that we've got enough buffer heads for flushing
the journal, the orignal code used __GFP_NOFAIL when performing
this allocation. Here we dispense with that in favour of using a
mempool. This should improve efficiency in low memory conditions
since flushing the journal is a good way to get memory back, we
don't want to be spinning, waiting on memory allocations. The
buffers which are allocated via this mempool are fairly short lived,
so that we'll recycle them pretty quickly.

Although there are other memory allocations which occur during the
journal flush process, this is the one which can potentially require
the most memory, so the most important one to fix.

The amount of memory reserved is a fixed amount, and we should not need
to scale it when there are a greater number of filesystems in use.

rhbz#796017
--
Signed-off-by: Bob Peterson <rpeterso@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>