dasd: fix fixpoint divide exception in define_extent
Patchwork-id: 46025
O-Subject: [RHEL6.3 PATCH 3/7] [s390x]dasd: fix fixpoint divide exception in define_extent
Bugzilla: 798002
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Description
-----------
Using a DASD with alias devices might trigger a BUG statement
with fixpoint divide exception.

If an IO request is build on an alias device without prefix enabled
we try to calculate with zero data from the alias device.
This can happen if the pathgroup is lost with an alias device already
in use.  Also the revalidation for re-established pathgroups is missing.

Prevent the alias device from being used with prefix not enabled.
Revalidate the server if a pathgroup is established.

Bugzilla
--------
BZ 798002
https://bugzilla.redhat.com/show_bug.cgi?id=798002

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.4
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f16330316321d1c388d13096f6858f5d2dac29dc
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=25e2cf1c1ac52d5078cf8cc3fd2f2ad084669ddd
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=b38f27e8425a132ed2dc49ffb3741404e81363d8
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=f9f8d02fae0dc47d8868fd069bb88d12f8d1d71f

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Jarod Wilson <jarod@redhat.com>