qeth: add missing wake_up call
Patchwork-id: 46028
O-Subject: [RHEL6.3 PATCH 6/7] [s390x]qeth: add missing wake_up call
Bugzilla: 806151
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Description
-----------
The cclengine hangs when running Communications Controller
for Linux on System z.

To send commands on the write channel 8 buffers exist.
If all 8 buffers are used, a wait is triggered on the
write channel. When such buffers are freed, a wake_up
is needed.

This patch adds the missing wake_up in qeth_release_buffer().
The fix is especially important when running Communications
Controller for Linux on System z.

Bugzilla
--------
BZ 806151
https://bugzilla.redhat.com/show_bug.cgi?id=806151

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.4
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=039055b965b9cfbeacb8cdcd1ef57e18b6374e58

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Jarod Wilson <jarod@redhat.com>