readahead: readahead page allocations are OK to fail
Patchwork-id: 46049
O-Subject: [RHEL6 PATCH] readahead: readahead page allocations are OK to fail
Bugzilla: 804815
RH-Acked-by: Dave Young <dyoung@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Jeff Garzik <jgarzik@redhat.com>

Hi,

This is a straight backport of the following:

  commit 7b1de5868b124d8f399d8791ed30a9b679d64d4d

    Pass __GFP_NORETRY|__GFP_NOWARN for readahead page allocations.

    readahead page allocations are completely optional.  They are OK to fail
    and in particular shall not trigger OOM on themselves.

LLNL apparently has a test case that triggers OOM kills in the readahead
path very consistently.  They've been testing with the above patch and
report that it solves the problem for them.  I've asked them to re-test
once the patch is in an official build.  I gave the patch a smoke test,
but didn't do any real targeted testing.

This resolves bug 804815.  Comments, as always, are appreciated.

Cheers,
Jeff

Signed-off-by: Jarod Wilson <jarod@redhat.com>