xen: only check xen_platform_pci_unplug if hvm
Patchwork-id: 46058
O-Subject: [RHEL6.3 PATCH] xen: only check xen_platform_pci_unplug if hvm
Bugzilla: 807354
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=807354
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=4199386

Commit 7b34e5d killed xenpv guest boots since they could no longer find
their root device. The problem is that with that patch we moved a condition
that used to be in an HVM only path to HVM _and_ PV paths. Thus the
condition should be 'if (HVM && <condition>)' as this patch makes it. Igor
has sent this patch upstream and it was warmly received, but it's still not
committed. I boot tested my brew build and it resolved the issue.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>