move is_vlan_dev into public header file
Patchwork-id: 46067
O-Subject: [PATCH RHEL6.3 1/2] net: move is_vlan_dev into public header file
Bugzilla: 713641
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

Users reported that LRO was not properly disabled when a VLAN device was
added to a bridge or routing was enabled.  This was already fixed
upstream (and in RHEL5) so this was an easy fix.

This patch is a backport of upstream commit:

	commit 6dcbbe25dcc9bd2bdeb4f685f8fb874ffc10e6be
	Author: Neil Horman <nhorman@tuxdriver.com>
	Date:   Tue May 24 08:31:08 2011 +0000

	    net: move is_vlan_dev into public header file (v2)

	    Migrate is_vlan_dev() to if_vlan.h so that core networkig can use it

I verified these two patches resolve the issue.

Brew build on 6.2.z confirms this will not break kabi and builds
successfully:

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4213628

They will resolve RHBZ 713641.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>