make dev_disable_lro use physical device if passed a vlan dev
Patchwork-id: 46068
O-Subject: [PATCH RHEL6.3 2/2] net: make dev_disable_lro use physical device if
	passed a vlan dev
Bugzilla: 713641
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

Users reported that LRO was not properly disabled when a VLAN device was
added to a bridge or routing was enabled.  This was already fixed
upstream (and in RHEL5) so this was an easy fix.

This patch is a backport of upstream commit:

commit f11970e383acd6f505f492f1bc07fb1a4d884829
Author: Neil Horman <nhorman@tuxdriver.com>
Date:   Tue May 24 08:31:09 2011 +0000

    net: make dev_disable_lro use physical device if passed a vlan dev (v2)

    If the device passed into dev_disable_lro is a vlan, then repoint the dev
    poniter so that we actually modify the underlying physical device.

I verified these two patches resolve the issue.

Brew build on 6.2.z confirms this will not break kabi and builds
successfully:

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4213628

They will resolve RHBZ 713641.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>