Fix race in process_vm_rw_core
Patchwork-id: 46170
O-Subject: [RHEL6.3 PATCH] mm: Fix race in process_vm_rw_core
Bugzilla: 739136
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=739136

mm_access is backported from the upstream series which added it. (It is
part of a larger series cleaning up /proc/$pid/mem, e268337d.) It'll be
a task for RHEL6.4 to backport the rest of the fixes...

Only upstream changed is that the cred_guard_mutex is in task_struct
instead of signal_struct in RHEL6.

commit 8cdb878dcb359fd1137e9abdee9322f5e9bcfdf8
Author: Christopher Yeoh <cyeoh@au1.ibm.com>
Date:   Thu Feb 2 11:34:09 2012 +1030

    Fix race in process_vm_rw_core

    This fixes the race in process_vm_core found by Oleg (see

      http://article.gmane.org/gmane.linux.kernel/1235667/

    for details).

    This has been updated since I last sent it as the creation of the new
    mm_access() function did almost exactly the same thing as parts of the
    previous version of this patch did.

    In order to use mm_access() even when /proc isn't enabled, we move it to
    kernel/fork.c where other related process mm access functions already
    are.

    Signed-off-by: Chris Yeoh <yeohc@au1.ibm.com>
    Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>