GFS2: Instruct DLM to avoid queue convert slowdowns
Patchwork-id: 46219
O-Subject: [RHEL6.3 PATCH BZ 799165 4/4] GFS2: Instruct DLM to avoid queue
	convert slowdowns
Bugzilla: 799165
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>
Date: Wed, 4 Apr 2012 15:14:51 -0500
Subject: [PATCH 4/4] GFS2: Instruct DLM to avoid queue convert slowdowns

This patch instructs DLM to prevent an "in place" conversion, where the
lock just stays on the granted queue, and instead forces the conversion to
the back of the convert queue. This is done on upward conversions only.

This is useful in cases where, for example, a lock is frequently needed in
PR on one node, but another node needs it temporarily in EX to update it.
This may happen, for example, when the rindex is being updated by gfs2_grow.
The gfs2_grow needs to have the lock in EX, but the other nodes need to
re-read it to retrieve the updates. The glock is already granted in PR on
the non-growing nodes, so this prevents them from continually re-granting
the lock in PR, and forces the EX from gfs2_grow to go through.

rhbz#799165

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>