xenfv: fix hangs when kdumping
Patchwork-id: 46227
O-Subject: [RHEL6.3 PATCH] xenfv: kdump: fix hangs when kdumping
Bugzilla: 811815
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=811815
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=4279876

Testing of kdump for Xen HVM guests found hangs on kexec. These
hangs were about 50% reproducible on some machines. qwan from
Xen QE identified that it happened anytime the message 'Grant
table initialized' was output to the console during the kexec
boot. Xen HVM guests have init code that unconditionally runs
on every boot to give it Xen-awareness. This code can lead to
reinitializing Xen components, such as the grant tables. When
kdumping we should avoid this, as we don't have the machinery
in place to handle it on the host nor in the guest. We can
avoid it by simply making this init conditional on not being
a kdump boot, which is what this patch does.

Upstream wouldn't want this, as it's a bit hacky. Fixing
kdump for pv-on-hvm is the right way to go there, but it
doesn't work upstream yet, so there's nothing we can backport
at the moment.

QE has tested the brew build and confirmed kdump now works
without hangs.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>