lpfc: Fix discovery problem when in pt2pt
Patchwork-id: 46245
O-Subject: [RHEL6.3 PATCH 17/36] scsi/lpfc: Fix discovery problem when in pt2pt
Bugzilla: 810522
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Rob Evers on behalf of Emulex <revers@redhat.com>

When a target is direct connected in pt2pt topology it is not discovered by the
driver. The confirm nport routine is called during plogi completion. It will
look up the ndlp using the service parameter wwpn in the target response. If the
ndlp that is returned in the lookup does not match the ndlp that the plogi was
sent with confirm_nport updates the new ndlp with old ndlp's information.
Confirm nport has to make sure that only one ndlp with that wwpn is active
before returning so it sets the old ndlp's state to NPR. It set the state
before it copied the state to the new ndlp so both ndlp's end up in NPR. When
the plogi completion routine calls the state machine with the plogi complete
event and the ndlp in NPR the ndlp'd state stays in NPR. The state machine is
stopped for this ndlp. Because it was the only target discovery is completed.
The old ndlp state is now copied to the new ndlp before setting the old one to
NPR

scsi-misc-2.6 commit id: eff4a01b6e9f8cee3c541ab7f2ad18b5bbffd124

https://bugzilla.redhat.com/show_bug.cgi?id=810522
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>