lpfc: create char device to take a reference
Patchwork-id: 46242
O-Subject: [RHEL6.3 PATCH 20/36] scsi/lpfc: create char device to take a
	reference
Bugzilla: 810522
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Rob Evers on behalf of Emulex <revers@redhat.com>

Fix management communication issues by creating character
device to take a reference on the driver

The management userspace applications have no way to take a
lock on the driver to prevent it from unloading. To remedy
this a character device is created that will up the reference
count on lpfc by one when it is opened and decrement it by
one when it is closed.

scsi-misc-2.6 commit id: 3ef6d24cd9f473518dd7941e86cc2a5f8992eed0

https://bugzilla.redhat.com/show_bug.cgi?id=810522
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>