Avoid check hlt for newer cpus
Patchwork-id: 46282
O-Subject: [RHEL6 PATCH] x86: Avoid check hlt for newer cpus
Bugzilla: 812439
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=812439

commit 2b107d93635616db0c3f893c8cc2e6d5cd8d77b2
Author: Jacob Pan <jacob.jun.pan@linux.intel.com>
Date:   Fri May 7 14:59:45 2010 -0700

    x86: Avoid check hlt for newer cpus

    Check hlt instruction was targeted for some older CPUs. It is an expensive
    operation in that it takes 4 ticks to break out the check.  We can avoid
    such check completely for newer x86 cpus (family >= 5).

    [ hpa: corrected family > 5 to family >= 5 ]

    Signed-off-by: Jacob Pan <jacob.jun.pan@linux.intel.com>
    LKML-Reference: <1273269585-14346-1-git-send-email-jacob.jun.pan@linux.intel.com>
    Signed-off-by: H. Peter Anvin <hpa@zytor.com>

My previous patch

	0a267f9 [x86] kdump: No need to disable ioapic in crash path

managed to break all 32-bit machines using kdump.  This was because on
there is a selftest called check_bugs.  Inside that function on
32-bit machines is a call to check_hlt which tests the halt instruction.
Well because my previous patch removed the disabling of the ioapic (which
cleaned up the ioapic and set it up for legacy mode), timer interrupts were
not getting through until later when the ioapic was actually setup correctly.
Therefore halt instructions would put the cpu to sleep and they would never
wake up.  Oops.

However, all this worked fine upstream.  Debugging led me to the above cheezy
commit.  Which just skips the check_hlt test if we are a Pentium Pro or higher.
Considering we don't support such old machines for RHEL-6, this commit should
be safe.

Tested on an amd box that PaulB. gave that was failing.  Kdump once again works
on 32-bit machines.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>