dm-raid: Record and handle missing devices.
Patchwork-id: 46396
O-Subject: [RHEL6.3 PATCH] DM RAID: Record and handle missing devices.
Bugzilla: 809231
RH-Acked-by: Milan Broz <mbroz@redhat.com>
RH-Acked-by: Alasdair Kergon <agk@redhat.com>
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>

Bug #809231
 https://bugzilla.redhat.com/show_bug.cgi?id=809231

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4312632

Testing:
 Testing shows issue is resolved.  (See bugzilla for more.)

Upstream status:
 Pending.  Neil Brown has queued them for linux-next (i.e. 3.5)
 https://www.redhat.com/archives/dm-devel/2012-April/msg00043.html

brassow

Missing dm-raid devices should be recorded in the superblock

When specifying the devices that compose a DM RAID array, it is possible to denote
failed or missing devices with '-'s.  When this occurs, we must record this in the
superblock.  We do this by checking if the array position's data device is missing
and then forcing MD to record the superblock by setting 'MD_CHANGE_DEVS' in
'raid_resume'.  If we do not cause the superblock to be rewritten by the resume
function, it is possible for a stale superblock to be written by an out-going
in-active table (during 'raid_dtr').

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>