default to clocksource unstable switching off
Patchwork-id: 46561
O-Subject: Re: [RHEL6 PATCH BZ 804535 3/3] default to clocksource unstable switching off
Bugzilla: 804535
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=804535

RHEL6 only.

The hpet & tsc are in synch for short periods of time (less than .5
seconds), however, they are not macroscopically in synch with one another.

If a long delay occurs (such as a sysrq-t) the clocksource will
erroneously switch from the tsc to the hpet because the clocksource cannot
properly calculate the time difference between the two clocks.

Disable the clocksource failover by default, and add a kernel parameter for
those who actually want it on.

A proper fix for this is still being discussed upstream (as well as some
other issues with the clocksource code).  BZ 809888 will introduce a
proper fix for this issue.

[v2]: changed parameter name to clocksource_failover (which is really what
is happening) and added additional details to !clocksource_failover
printk.

Signed-off-by: Jarod Wilson <jarod@redhat.com>