st: fix memory leak with 1MB tape I/O
Patchwork-id: 46631
O-Subject: [RHEL6.3 PATCH] scsi: memory leak with 1MB tape I/O
Bugzilla: 811703
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

GSS has multiple customers hitting this problem, David Jeffery
has posted a fix upstream and it has been ACK'd by st driver
maintainer.

http://marc.info/?l=linux-scsi&m=133416027105522&w=2

There is a memory leak in the st driver when sending large enough reads or
writes using st's direct I/O path.  As part of mapping the application's
memory, a buffer to hold page pointers is allocated and the count of mapped
pages is stored in field do_dio.  A non-zero do_dio marks that direct I/O is
in use.

But do_dio is only 1 byte in size.  Mapping 256 4k pages overflows
do_dio and causes it to be set to 0, like direct I/O option was not
used.  When the I/O completes, the buffer to hold the page pointers is
not freed, and the page counts of the mapped pages are not reduced.
Every I/O of this size then leaks memory.

The size of do_dio needs to be increased to prevent it wrapping around.

signed-off-by: David Jeffery <djeffery@redhat.com>

https://brewweb.devel.redhat.com/taskinfo?taskID=4334958

Patch has been verified by customer, and resolves BZ 811703

Signed-off-by: Jarod Wilson <jarod@redhat.com>