ixgbe: Correct Adaptive Interrupt Moderation so that it will change values
Patchwork-id: 46698
O-Subject: [PATCH RHEL6.3] ixgbe: Correct Adaptive Interrupt Moderation so that it will change values
Bugzilla: 802837
RH-Acked-by: Jesper Brouer <jbrouer@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Adam Okuliar reported a performance regression on the latest ixgbe code
compared to RHEL6.2.  It was tough to track as the throughput as well as
CPU utilization needed to be monitored to detect which patch was causing
the problem.

It turned out this was fixed shortly after we stopped taking patches
from upstream, so that was why we missed it.  The description also does
not specifically indicate what not modifying the adaptive interrupt
moderation value would do to the system, so that wasn't helpful.

This is a backport of the following upstream commit:

commit 5d967eb7f84d2761d624d033c2583187b37b56ce
Author: Alexander Duyck <alexander.h.duyck@intel.com>
Date:   Wed Feb 8 07:49:43 2012 +0000

    ixgbe: Correct Adaptive Interrupt Moderation so that it will change values

I verified this will resolve the problem reported in RHBZ 802837 and Tom
Tracy verified this resolves an reported (but not in bugzilla) issue
related to transmit latency.

Signed-off-by: Jarod Wilson <jarod@redhat.com>