vfs: fix LOOKUP_DIRECTORY not propagated to managed_dentry()
Patchwork-id: 46734
O-Subject: [RHEL-6.3 PATCH 1/2] vfs - fix LOOKUP_DIRECTORY not propagated to managed_dentry()
Bugzilla: 786149
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=786149

The vfs-automount infrastructure assumes that LOOKUP_DIRECTORY will
be included in the nameidata flags if a trailing slash is given on
the path being walked. But this flag is private to __link_path_walk()
so it must be added when looking up the last component.

Also if the path being walked must follow a symlink that ends with a
"/" the nameidata flag LOOKUP_DIRECTORY can be set earlier than the
last path component and lead to an incorrect ENOTDIR.  For the purpose
of possible automounting this flag isn't needed for intermediate path
components because the LOOKUP_CONTINUE flag will be set. So the check
must be made in do_lookup() so the LOOKUP_DIRECTORY flag is set only
for the required call to follow_managed().

Signed-off-by: Ian Kent <ikent@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>