cifs: check S_AUTOMOUNT in revalidate
Patchwork-id: 46735
O-Subject: [RHEL-6.3 PATCH 2/2] cifs - check S_AUTOMOUNT in revalidate
Bugzilla: 786149
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=786149

When revalidating a dentry, if the inode wasn't known to be a dfs
entry when the dentry was instantiated, such as when created via
->readdir(), the DCACHE_NEED_AUTOMOUNT flag needs to be set on the
dentry in ->d_revalidate().

The false return from cifs_d_revalidate(), due to the inode now
being marked with the S_AUTOMOUNT flag might not invalidate the
dentry if there is a concurrent path walk. This is because the
dentry reference count will be at least 2 in this case causing
d_invalidate() to return EBUSY. So the asumption that the dentry
will be discarded then correctly instantiated via ->lookup() might
not hold.

Signed-off-by: Ian Kent <ikent@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>