mqueue: strengthen checks on mqueue creation
Patchwork-id: 46756
O-Subject: [Patch RHEL6 3/8] ipc/mqueue: strengthen checks on mqueue creation
Bugzilla: 750260
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>

We already check the mq attr struct if it's passed in, but now that the
admin can set system wide defaults separate from maximums, it's actually
possible to set the defaults to something that would overflow.  So,
if there is no attr struct passed in to the open call, check the default
values.

While we are at it, simplify mq_attr_ok() by making it return 0 or an
error condition, so that way if we add more tests to it later, we have
the option of what error should be returned instead of the calling
location having to pick a possibly inaccurate error code.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit f2c33aff95ebefcca9e5dc3c222134a5b44f21aa)

Conflicts:

	ipc/mqueue.c

Minor difference in how code was factored between upstream and our
version, but logically was the same.

Signed-off-by: Jarod Wilson <jarod@redhat.com>